Remoção de interface de service e remoção de lombok do projeto
